feat: Validator class를 통해 유효성 검사를 진행하는 기능 구현
